requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
whitenoise==6.6.0
